Armor bonus update
Updated the basic Armor data set
Updated the Armor Display to show the bonus as a raw number... 2/1(1)
instead of 2/3(1). That way I can show the cumulative bonus in the
Defense Information Display in a way it makes sense. If you have two
items that both have a bonus absorb it will show... 3/2(2) instead of
3/5(2) and you know you only have 2 rolls to make.