Spells
Moved the tables to SpellList
Made SpellListDisplay use the CardLayout to switch between the different
SpellList's the player may have... (might want to build a consolidated
spell list of all spells - maybe on the blank page when "Select magic
Area" is selected.)

empty 'cards' are reused for the newly selected spell area and only
saved once a spell has been added to it.

Still need to be able to remove spells.
Some descriptions are still wonky.