Determination Points - Still under construction.
A lot more work on Determination Points

Enable/Disable as appropriate.
Added toString for debug.
Moved the "Learn" and "Give Up" buttons to their individual panels.
Moved the button actions to the individual panels.
Moved the description panel to the top of the page.
Added the DPCost to the magic spell panel based on spell level.

Still have debugging code in the Determination Points Page.